Ref: Replace string to uuid in validators.

+++ solution/microservices/ad_engine/src/domain/schemas/request/time.rs
@@ -5,5 +5,6 @@
 )]
 pub struct TimeAdvanceRequest {
     #[schema(example = 1)]
-    pub current_date: usize,
+    #[validate(range(min = 0, message = "date must be under or equal 0"))]
+    pub current_date: u16,
 }
